CBL-7394 : Add mdns logging domain
* Added mdns logging domain to both objective-c and swift

* Reserved 128 and 256 value for bluetooth and wifi-aware that we may have in the future for multipeer replicator.

* Changed Swift LogDomain type from UInt8 to UInt.

* Fixed missing switch cases in CBLFileLogSink.mm.

+++ Swift/LogSinks.swift
@@ -75,7 +75,7 @@ public class LogSinks {
         
         func writeLog(with level: CBLLogLevel, domain: CBLLogDomain, message: String) {
             let logLevel = LogLevel.init(rawValue: UInt8(level.rawValue))!
-            let logDomain = LogDomain.init(rawValue: UInt8(domain.rawValue))!
+            let logDomain = LogDomain.init(rawValue: domain.rawValue)!
             logSink.writeLog(level:logLevel, domain: logDomain, message: message)
         }
     }
